+++ test/compute-engine/latex-syntax/supsub.test.ts
@@ -97,7 +97,7 @@ describe('SUPSUB', () => {
           "Power",
           [
             "Error",
-            ["ErrorCode", "'incompatible-domain'", "Numbers", "Symbols"],
+            ["ErrorCode", "'incompatible-domain'", "Numbers", "Anything"],
             ["Subscript", "x", ["Add", "r", 1]]
           ],
           ["Add", "s", 1]
@@ -111,7 +111,7 @@ describe('SUPSUB', () => {
           "Power",
           [
             "Error",
-            ["ErrorCode", "'incompatible-domain'", "Numbers", "Symbols"],
+            ["ErrorCode", "'incompatible-domain'", "Numbers", "Anything"],
             ["Subscript", "x", ["Add", "p", 1]]
           ],
           ["Add", "q", 1]
@@ -120,7 +120,7 @@ describe('SUPSUB', () => {
           "Power",
           [
             "Error",
-            ["ErrorCode", "'incompatible-domain'", "Numbers", "Symbols"],
+            ["ErrorCode", "'incompatible-domain'", "Numbers", "Anything"],
             ["Subscript", "x", ["Add", "r", 1]]
           ],
           ["Add", "s", 1]
@@ -135,9 +135,17 @@ describe('SUPSUB', () => {
     expect(parse('(x+1)_{n-1}')).toMatchInlineSnapshot(
       `["Subscript", ["Add", "x", 1], ["Subtract", "n", 1]]`
     );
-    expect(parse('(x+1)^n_0')).toMatchInlineSnapshot(
-      `["Power", ["Subscript", ["Add", "x", 1], 0], "n"]`
-    );
+    expect(parse('(x+1)^n_0')).toMatchInlineSnapshot(`
+      [
+        "Power",
+        [
+          "Error",
+          ["ErrorCode", "'incompatible-domain'", "Numbers", "Anything"],
+          ["Subscript", ["Add", "x", 1], 0]
+        ],
+        "n"
+      ]
+    `);
     expect(parse('^p_q{x+1}^n_0')).toMatchInlineSnapshot(`
       [
         "Superscript",